由于axios接收long类型丢失精度 所以改为int型

+++ src/main/java/com/zz/pojo/Community.java
@@ -5,12 +5,11 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 
 @Data
 public class Community implements Serializable {
     @TableId(value = "community_id", type = IdType.ASSIGN_ID)
-    private BigInteger communityId;
+    private Integer communityId;
     private String communityName;
     private String communityInfo;
 }
