+++ src/main/java/com/zz/pojo/ServiceComment.java
@@ -5,16 +5,15 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 
 @Data
 public class ServiceComment implements Serializable {
     @TableId(value = "comment_id", type = IdType.ASSIGN_ID)
-    private BigInteger commentId;
-    private BigInteger userId;
+    private Integer commentId;
+    private Integer userId;
     private String context;
     private String date;
     private Integer commentLevel;
-    private BigInteger serviceOrderId;
-    private BigInteger houseKeeperId;
+    private Integer serviceOrderId;
+    private Integer houseKeeperId;
 }
