+++ src/main/java/com/zz/pojo/User.java
@@ -5,12 +5,11 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 
 @Data
 public class User implements Serializable {
      @TableId(value = "user_id", type = IdType.ASSIGN_ID)
-     private BigInteger userId;
+     private Integer userId;
      private String userName;
      private String password;
      private String phone;
@@ -18,6 +17,6 @@ public class User implements Serializable {
      private Sex sex;
      private String cardId;
      private String address;
-     private BigInteger communityId;
+     private Integer communityId;
 
 }
